math-tutor-backend/requirements.txt
# --- Web Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0

# --- AI & LLM ---
openai==1.12.0
python-dotenv==1.0.1

# --- Security & Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# --- Database (SQLAlchemy) ---
sqlalchemy==2.0.25

# --- Utilities ---
requests==2.31.0
pydantic[email]==2.6.1